requirements.txt
asgiref==3.11.1
Django==6.0.2
django-debug-toolbar==6.2.0
django-extensions==4.1
pillow==12.1.1
sqlparse==0.5.5
django-thumbnails==0.8.0
django-ckeditor==6.7.3